WaypointInserter will add more WPT on you the Route to your GPX file
If your GPX file doesn't contain enough WPTs in it - your navigation via xCode project could be "jumpy", "glitchy" and very fast. It happen because lack of Waypoints on your route. To smooth your navigation on the route - you should add more WPTs in your GPX file. Making it manually is a very long and painful process, so WaypointInserter will do it in 2 clicks for you.